Added SoftDealer and simulator
Soft dealers will hit on 17 if an ace is counted as an 11, and never
hit/double.

botGames function allows us to decide to use one bot as a player and
another bot as a dealer. it will report the total games won/played of
the player.

Will add additional information to report and work on outputting the
information to an easier to read format.