Added BasicPlayer
BasicPlayer plays a game of Blackjack optimized for a hard dealer,
although should be comparable for a soft dealer as well.

Player makes decisions based on their hand configuration and the
dealer's one face-up card. These values form the basis of expert optimal
strategy in blackjack, and should serve as a benchmark to compare the
Monte Carlo Tree Search algorithm